uds: Copy ownership of server socket in pmc clients.
ptp4l sending a response to a pmc client needs to have permissions to
write to the client's UNIX domain socket. If ptp4l runs under a non-root
user, it cannot write to sockets bound by the pmc client if it did that
as root.

After binding the client socket, change its owner to the owner of the
server socket, so it can send the client a response.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>