Follow server UDS location in pmc clients.
Modify pmc, phc2sys, ts2phc, and tz2alt to bind their client sockets in
the same directory as the server's socket instead of hardcoded /var/run.
In the pmc program it's the default, it can still be configured with the
-i option to any address.

This allows the clients to be started under the same user as to which
ptp4l is configured to switch, with all their sockets in the same
directory where the user has write permissions. phc2sys and ts2phc will
need also permissions to open the PHC devices and the CAP_SYS_TIME
capability needed to adjust the clock (e.g. set by setpriv(1)).

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>
Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>