servo: Add option to enable/disable servo.
Add servo_enabled configuration option to allow the servo to be
disabled. When disabled, the servo will stay in the UNLOCKED state and
the port in the UNCALIBRATED state.

This value will be configurable at runtime and will allow switching the
clock control between multiple active ptp4l instances, e.g. as is
required in the G.8265.1 profile.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>