sysoff: Add support for auxiliary system clocks.
Add a new parameter to the probe and measure functions to select the
system clock for the PTP_SYS_OFFSET* ioctls.

Currently, only the PTP_SYS_OFFSET_EXTENDED ioctl can use a different
clock than CLOCK_REALTIME and actually supports CLOCK_AUX. The clock is
specified in one of the original reserved fields. Detect if the
installed kernel header already has the renamed field, and if not, set
the clock ID in the reserved field.

Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>