phc2sys: Add support for auxiliary system clocks.
Update paths specific to CLOCK_REALTIME to cover also CLOCK_AUX* clocks.

Add -C option to specify which system clock should be used with in the
automatic mode (-a and -r options). Only one system clock can be used in
the automatic mode.

In the static mode, CLOCK_AUX* clocks can be specified by the -c and
-s options.

To allow operation with multiple system clocks, delay probing of the
supported PHC sysoff method until it is needed and the selected system
clock is known. Track the sysoff clock ID for each PHC and probe it
again if the system clock is changed. If multiple system clocks are
synchronized to the same PHC, the probe will repeat with each clock
update. To avoid that, the probed methods would need to be cached,
potentially for all (PHC, sysclock) pairs.

Reviewed-by: Jacob Keller <jacob.e.keller@intel.com>
Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>