timemaster: Add option to merge PTP interfaces.
Add merge option to the PTP domain section to enable or disable a
preference to merge interfaces of a domain into a single ptp4l instance
in order to provide a single source to the NTP implementation and leave
the selection of sources in the domain to ptp4l.

For interfaces using HW timestamping with different PHCs set the
boundary_clock_jbod option of ptp4l and switch to the new -e option of
phc2sys to actually have the PHCs synchronized to each other.

Signed-off-by: Miroslav Lichvar <mlichvar@redhat.com>